+++ BE/app/src/main/java/com/todo/app/common/config/FilterConfig.java
@@ -1,15 +1,15 @@
 package com.todo.app.common.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.todo.app.common.aop.jwt.CorsFilter;
-import com.todo.app.common.aop.jwt.JwtAuthorizationFilter;
+import com.todo.app.common.filter.CorsFilter;
+import com.todo.app.common.filter.JwtAuthorizationFilter;
 import javax.servlet.Filter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class JwtConfig {
+public class FilterConfig {
 
     @Bean
     public FilterRegistrationBean<Filter> corsFilter() {
